# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth · Login & Protect API",
    "version": "1.0.0",
    "description": "Express REST API with Supabase Auth and JWT Guarded Routes"
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "test@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing credentials" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user and return tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "test@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public endpoint",
        "responses": {
          "200": { "description": "Public info retrieved" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get protected user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Profile retrieved" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}